# openapi.yaml
openapi: 3.1.0
info:
  title: Event Ticketing API
  version: "1.3.0-draft"
  description: >
    Full v1 marketplace surface for a multi-organizer event ticketing
    platform: identity/auth, organization onboarding (application +
    admin approval) and org-scoped roles, events & ticket types, cart &
    checkout (with promo codes), orders & tickets, ticket transfer &
    organizer-controlled resale, waitlists, refunds & payouts, ticket
    templates, check-in/scanning, notifications, disputes, admin
    moderation actions, audit log, and analytics. Aligned to requirements.md v1.0, business-rules.md v1.0,
    and use-cases.md v1.0. Genuinely open/undecided items (target
    markets, resale opt-in/out default, mid-event check-in mode
    switching, etc.) are tracked in requirements.md section 7, not
    modeled here.
servers:
  - url: https://api.example.com/v1
    description: Placeholder — actual host TBD in the architecture phase.
security:
  - bearerAuth: []
...
paths:
...
  /organizations/{orgId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: getOrganization
      tags: [Organizations]
      summary: Get an Organization (the applicant, its members, or admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateOrganization
      tags: [Organizations]
      summary: Update an Organization (owner or organizer only; approved orgs only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrganizationUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
...
  /platform-fees:
    get:
      operationId: listPlatformFeeRules
      tags: [PlatformFees]
      summary: List the platform fee rules (admin only)
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [PLATFORM, ORGANIZATION, EVENT] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlatformFeeRule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /platform-fees/effective:
    get:
      operationId: getEffectivePlatformFee
      tags: [PlatformFees]
      summary: Which rule applies to an event right now (admin only)
      parameters:
        - name: eventId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK - `rule` is null when no rule applies (no fee is charged)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EffectivePlatformFee" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /platform-fees/default:
    put:
      operationId: setDefaultPlatformFee
      tags: [PlatformFees]
      summary: Set (or replace) the platform-wide default fee (admin only)
      description: |
        The fee ("admin cut") added on top of the ticket price of every purchase that has no more specific rule.
        A percentage is worked out on the ticket total after promo discounts; a flat amount is per order.
        A rate of 0 is allowed and means no fee.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformFeeRuleRequest" }
      responses:
        "200":
          description: The rule now in force
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformFeeRule" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      operationId: removeDefaultPlatformFee
      tags: [PlatformFees]
      summary: Remove the platform-wide default (admin only) - purchases without an override then pay no fee
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /platform-fees/organizations/{organizationId}:
    parameters:
      - name: organizationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: setOrganizationPlatformFee
      tags: [PlatformFees]
      summary: Override the fee for one organization (admin only) - applies to its events that have no event rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformFeeRuleRequest" }
      responses:
        "200":
          description: The rule now in force
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformFeeRule" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: removeOrganizationPlatformFee
      tags: [PlatformFees]
      summary: Remove an organization's override (admin only) - the platform default applies again
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /platform-fees/events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: setEventPlatformFee
      tags: [PlatformFees]
      summary: Override the fee for one event (admin only) - the most specific rule, it beats the organization and the default
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformFeeRuleRequest" }
      responses:
        "200":
          description: The rule now in force
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformFeeRule" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: removeEventPlatformFee
      tags: [PlatformFees]
      summary: Remove an event's override (admin only) - the organization's rule, else the default, applies again
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /events/{eventId}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      operationId: getEvent
      tags: [Events]
      security: []
      summary: Get an Event (public if published; organizer/owner/admin otherwise)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateEvent
      tags: [Events]
      summary: Update an Event (organizer or owner of the owning organization only - an admin gets 403)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
...
  /events/{eventId}/scanner-devices:
    parameters:
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: authorizeScannerDevice
      tags: [CheckIn]
      summary: Authorize a new ScannerDevice for an event (owning organizer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_label: { type: string }
                force_replace:
                  type: boolean
                  default: false
                  description: Required to replace the sole device in pure_offline mode (the override flow, requirements 4.11).
      responses:
        "201":
          description: Authorized — credential is shown once, here, and never again
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScannerDeviceWithCredential" }
        "409":
          description: pure_offline mode already has an active device; retry with force_replace true
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /scanner-devices/{deviceId}:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    delete:
      operationId: revokeScannerDevice
      tags: [CheckIn]
      summary: Revoke a device's credential immediately (owning organizer)
      responses:
        "204": { description: Revoked }
  /scanner-devices/{deviceId}/dataset:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    get:
      operationId: getScannerDeviceDataset
      tags: [CheckIn]
      security:
        - deviceAuth: []
      summary: Pre-fetch the ticket dataset for offline validation
      responses:
        "200":
          description: >
            Full event ticket dataset. In pure_offline mode this has no
            expiry for the sole device; in standard mode the client must
            treat it as stale after CheckInConfig.offline_fallback_expiry_seconds.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketDataset" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceAuth:
      type: http
      scheme: bearer
      description: A ScannerDevice's own credential, issued by POST /events/{eventId}/scanner-devices.
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated, but not permitted (wrong role/scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: integer, description: Minor units (e.g. cents) }
        currency: { type: string, description: ISO 4217, example: USD }
...
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected]
          description: An application in `pending` grants no organization privileges. Only `approved` gains an owner and can publish events or receive payouts.
        documents:
          type: array
          description: Verification documents submitted with the application (e.g. business permit).
          items:
            type: object
            properties:
              type: { type: string, example: business_permit }
              url: { type: string, format: uri }
        owner_id:
          type: [string, "null"]
          format: uuid
          description: Set only once the application is approved; equals the original applicant (created_by).
        created_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, description: The applicant who submitted this organization's application. }
...
    OrganizationUpdate:
      type: object
      properties:
        name: { type: string }
...
    Event:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        venue:
          type: [object, "null"]
          description: Snapshot of the referenced Venue at read time, for convenience; null for a fully virtual event with no venue_id.
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            address: { type: string }
            latitude: { type: [number, "null"] }
            longitude: { type: [number, "null"] }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        timezone: { type: string, example: Asia/Manila }
        images:
          type: array
          items: { type: string, format: uri }
        status: { type: string, enum: [draft, published, on_sale, sold_out, cancelled, completed] }
        ticket_number_prefix:
          type: string
          minLength: 3
          maxLength: 3
          description: Random 3-letter (A-Z) code, reserved unique platform-wide at creation; seeds Ticket.ticket_number only.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, description: User (organization member) who created the event. }
        updated_by: { type: string, format: uuid, description: User who last modified the event. }
...
    EventUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        images:
          type: array
          items: { type: string, format: uri }
...
    TicketDataset:
      type: object
      description: Full pre-fetched ticket state for offline validation.
      properties:
        event_id: { type: string, format: uuid }
        generated_at: { type: string, format: date-time }
        tickets:
          type: array
          items:
            type: object
            properties:
              ticket_id: { type: string, format: uuid }
              credential_hash: { type: string, description: For local signature/lookup verification without exposing the raw credential unnecessarily. }
              status: { type: string, enum: [valid, used, transferred, refunded, cancelled] }
...
    ScannerDeviceWithCredential:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        device_label: { type: string }
        status: { type: string, enum: [active, revoked] }
        credential:
          type: string
          description: Shown once, at creation, and never again.
        created_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, description: Organizer who authorized this device. }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string, format: uuid }
...
    PlatformFeeRuleRequest:
      type: object
      required: [type]
      description: PERCENTAGE needs `percentage` and no `flatAmount`; FLAT needs `flatAmount` and no `percentage`.
      properties:
        type: { type: string, enum: [PERCENTAGE, FLAT] }
        percentage: { type: number, minimum: 0, maximum: 100, multipleOf: 0.01, description: "Percent of the ticket total after promo discounts, rounded half up to the minor unit. 0 = no fee." }
        flatAmount:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: A fixed fee per order. A flat rule in another currency than the order is skipped (no fee).
    PlatformFeeRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        scope: { type: string, enum: [PLATFORM, ORGANIZATION, EVENT], description: "The most specific scope wins: EVENT, then ORGANIZATION, then PLATFORM." }
        scopeId: { type: [string, "null"], format: uuid, description: "The organization or event id; null for the platform default." }
        type: { type: string, enum: [PERCENTAGE, FLAT] }
        percentage: { type: [number, "null"] }
        flatAmount:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: [string, "null"], format: date-time }
    EffectivePlatformFee:
      type: object
      properties:
        eventId: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        rule:
          oneOf:
            - $ref: "#/components/schemas/PlatformFeeRule"
            - type: "null"
          description: The rule that applies to the event; null when none does (no fee).
...